break up mgmt into smaller pieces
